feat: show error on screen
