+++ lib/features/sks-menu/presentation/sks_menu_screen.dart
@@ -3,8 +3,12 @@ import "dart:core";
 import "package:auto_route/annotations.dart";
 import "package:flutter/material.dart";
 import "package:flutter_riverpod/flutter_riverpod.dart";
+import "package:logger/logger.dart";
+import "package:lottie/lottie.dart";
 
+import "../../../../theme/app_theme.dart";
 import "../../../config/ui_config.dart";
+import "../../../gen/assets.gen.dart";
 import "../../../utils/context_extensions.dart";
 import "../../../widgets/detail_views/detail_view_app_bar.dart";
 import "../../home_view/widgets/paddings.dart";
@@ -13,7 +17,6 @@ import "../data/models/sks_menu_response.dart";
 import "../data/repository/sks_menu_repository.dart";
 import "widgets/sks_menu_data_source_link.dart";
 import "widgets/sks_menu_header.dart";
-import "widgets/sks_menu_loading.dart";
 import "widgets/sks_menu_section.dart";
 
 @RoutePage()
@@ -38,7 +41,7 @@ class SksMenuView extends ConsumerWidget {
             ),
         appBarPopTitle,
       ),
-      error: (error, stackTrace) => const SKSMenuLottieAnimation(),
+      error: (error, stackTrace) => _SKSMenuLottieAnimation(error: error),
       loading: () => const Scaffold(
         body: Center(
           child: CircularProgressIndicator(),
@@ -55,8 +58,8 @@ class _SksMenuView extends StatelessWidget {
   final String? appBarPopTitle;
   @override
   Widget build(BuildContext context) {
-    if (sksMenuData.meals.isEmpty) {
-      return const SKSMenuLottieAnimation();
+    if (!sksMenuData.isMenuOnline) {
+      return const _SKSMenuLottieAnimation();
     }
     return Scaffold(
       appBar: DetailViewAppBar(
@@ -82,6 +85,51 @@ class _SksMenuView extends StatelessWidget {
           const SizedBox(
             height: ScienceClubsViewConfig.mediumPadding,
           ),
+        ],
+      ),
+    );
+  }
+}
+
+class _SKSMenuLottieAnimation extends StatelessWidget {
+  const _SKSMenuLottieAnimation({
+    this.error,
+  });
+
+  final Object? error;
+  @override
+  Widget build(BuildContext context) {
+    Logger().e(error.toString());
+    return Scaffold(
+      body: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          SizedBox.square(
+            dimension: 200,
+            child: Lottie.asset(
+              Assets.animations.sksClosed,
+              fit: BoxFit.cover,
+              repeat: false,
+              frameRate: const FrameRate(LottieAnimationConfig.frameRate),
+              renderCache: RenderCache.drawingCommands,
+            ),
+          ),
+          Align(
+            child: Text(
+              context.localize.sks_menu_closed,
+              style: context.textTheme.headline,
+              textAlign: TextAlign.center,
+            ),
+          ),
+          if (error != null)
+            Padding(
+              padding: const EdgeInsets.only(top: 8),
+              child: Text(
+                error.toString(),
+                style: context.textTheme.titleGrey,
+                textAlign: TextAlign.center,
+              ),
+            ),
         ],
       ),
     );
